move choice to new class RoadProbability to improve complexity
