+++ Assets/Scripts/DataTypes/RouteProbabilities.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 1a2e96ae57e0f094288bc51b52f58995
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
